Refactor application banner component
