+++ apps/web/src/app/(main)/[atUserName]/(details)/layout.tsx
@@ -1,6 +1,6 @@
 import * as styles from './layout.css';
-import { ApplicationBanner } from '../../../../components/domains/application/application-banner';
 import { ApplicationLayout } from '../../../../components/domains/application/application-layout';
+import { Banner } from '../../../../components/elements/banner';
 import { createMetadata } from '../../../../helpers/create-metadata';
 import { findUserByName } from '../../../../repositories/user-repository';
 import { getUserName } from '../_helpers/get-user-name';
@@ -39,7 +39,7 @@ const UserDetailsLayout: FC<UserDetailsLayoutProps> = ({
   //TODO: Make it retrieve the user's mute status.
   return (
     <ApplicationLayout header={header}>
-      <ApplicationBanner>This user is currently muted.</ApplicationBanner>
+      <Banner>This user is currently muted.</Banner>
       <main className={styles.main}>
         <article className={styles.article}>
           {profile}
