--- apps/web/src/components/domains/application/application-banner/index.ts
@@ -1,1 +1,0 @@
-export * from './application-banner';
