+++ apps/web/src/components/elements/banner/banner.tsx
@@ -1,12 +1,12 @@
 import { clsx } from 'clsx';
 
-import * as styles from './application-banner.css';
+import * as styles from './banner.css';
 
 import type { ComponentPropsWithoutRef, FC } from 'react';
 
-export type ApplicationBannerProps = ComponentPropsWithoutRef<'div'>;
+export type ApplicationBannerProps = ComponentPropsWithoutRef<'aside'>;
 
-export const ApplicationBanner: FC<ApplicationBannerProps> = ({
+export const Banner: FC<ApplicationBannerProps> = ({
   className,
   children,
   ...props
