+++ apps/web/src/components/elements/banner/index.ts
@@ -1,0 +1,1 @@
+export * from './banner';
